refactor(mobile): remove unused PR and review code
