+++ mobile/src/session/mobile-diff-review-state.ts
@@ -104,31 +104,6 @@ export function mergeMobileDiffReviewState(
   }
 }
 
-export function markMobileDiffReviewFileOpened(
-  state: MobileDiffReviewState,
-  descriptor: MobileDiffReviewFileDescriptor,
-  now: number
-): MobileDiffReviewState {
-  const previous = state.files[descriptor.key]
-  return {
-    ...state,
-    updatedAt: now,
-    files: {
-      ...state.files,
-      [descriptor.key]: {
-        key: descriptor.key,
-        filePath: descriptor.filePath,
-        oldPath: descriptor.oldPath,
-        scope: descriptor.scope,
-        reviewedAt: previous?.reviewedAt,
-        reviewDiffIdentity: previous?.reviewDiffIdentity,
-        lastOpenedAt: now,
-        lastSeenDiffIdentity: descriptor.diffIdentity
-      }
-    }
-  }
-}
-
 export function markMobileDiffReviewFileReviewed(
   state: MobileDiffReviewState,
   descriptor: MobileDiffReviewFileDescriptor,
