+++ mobile/src/session/use-native-chat-action-outcomes.ts
@@ -1,5 +1,4 @@
 import { useCallback } from 'react'
-import type { MobileNativeChatSendOutcome } from './mobile-native-chat-send'
 
 /** Wraps a chat card action so an accepted write also retires the route's held
  *  failure banner.
@@ -21,17 +20,5 @@ export function useNativeChatAcceptedAction<Params extends unknown[]>(
       return accepted
     },
     [action, onAccepted]
-  )
-}
-
-/** Boolean surface for callers with no pre-pasted input: 'unknown' stays true
- *  (the send usually landed; the optimistic echo is already held unconfirmed). */
-export function useNativeChatSentFlag(
-  send: (text: string, images?: string[]) => Promise<MobileNativeChatSendOutcome>
-): (text: string, images?: string[]) => Promise<boolean> {
-  return useCallback(
-    async (text: string, images?: string[]): Promise<boolean> =>
-      (await send(text, images)) !== 'rejected',
-    [send]
   )
 }
