+++ mobile/src/source-control/mobile-open-pr-prefill.ts
@@ -1,14 +1,7 @@
-import type { RpcClient } from '../transport/rpc-client'
 import { readMobileGitStatusResult } from '../session/mobile-diff-review-rpc'
 import type { MobileGitStatusResult } from './mobile-git-status'
-import { resolveMobilePrPrefill, type MobilePrPrefill } from './mobile-pr-create'
 
-// Resolves the create-PR prefill from a git status snapshot. Split from the
-// runners hook to keep that file under the line limit.
-
-// Reads a fresh git.status after a push so the prefill reflects the just-pushed
-// branch's upstream/ahead data instead of the pre-push captured status. Best-effort:
-// returns the captured status on any read failure.
+// Refresh after a push when possible so readiness reflects the new upstream state.
 export async function readFreshGitStatus(
   worktreeId: string,
   fallback: MobileGitStatusResult | null,
@@ -20,23 +13,6 @@ export async function readFreshGitStatus(
   } catch {
     return fallback
   }
-}
-
-export async function buildOpenPrPrefill(
-  client: Pick<RpcClient, 'sendRequest'> | null,
-  worktreeId: string,
-  status: MobileGitStatusResult | null,
-  branchLabel: string
-): Promise<MobilePrPrefill> {
-  if (!client) {
-    return { provider: 'github', base: 'main', title: branchLabel, body: '' }
-  }
-  const gitReadiness = getMobilePrEligibilityReadiness(status)
-  return resolveMobilePrPrefill(client, worktreeId, {
-    branch: status?.branch,
-    title: branchLabel,
-    ...gitReadiness
-  })
 }
 
 export function getMobilePrEligibilityReadiness(status: MobileGitStatusResult | null): {
